requirements.txt
torch>=2.0.0
sentence-transformers>=2.6.0
transformers>=4.44.0
numpy>=1.24.0
qdrant-client
streamlit
python-docx
rank_bm25
pyvi
scikit-learn
google-generativeai
